pyproject.toml
[project]
name = "grok2api"
version = "0.1.0"
description = "Grok2API (FastAPI backend + Cloudflare Workers/Pages variant)"
readme = "readme.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiomysql>=0.2.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.12.3",
    "curl-cffi>=0.13.0",
    "fastapi>=0.119.0",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "orjson>=3.11.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "requests>=2.32.3",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.46",
    "tomli>=2.4.0",
    "uvicorn>=0.37.0",
]
